Rename old and new packet packages
