+++ src/main/java/org/ethereum/beacon/discovery/network/NetworkParcel.java
@@ -5,8 +5,8 @@
 package org.ethereum.beacon.discovery.network;
 
 import java.net.InetSocketAddress;
-import org.ethereum.beacon.discovery.packet5_1.Packet;
-import org.ethereum.beacon.discovery.packet5_1.RawPacket;
+import org.ethereum.beacon.discovery.packet.Packet;
+import org.ethereum.beacon.discovery.packet.RawPacket;
 
 /**
  * Abstraction on the top of the {@link Packet}.
