+++ src/main/java/org/ethereum/beacon/discovery/packet/impl/HeaderImpl.java
@@ -1,16 +1,16 @@
-package org.ethereum.beacon.discovery.packet5_1.impl;
+package org.ethereum.beacon.discovery.packet.impl;
 
 import static com.google.common.base.Preconditions.checkArgument;
 
 import javax.crypto.Cipher;
 import org.apache.tuweni.bytes.Bytes;
-import org.ethereum.beacon.discovery.packet5_1.AuthData;
-import org.ethereum.beacon.discovery.packet5_1.DecodeException;
-import org.ethereum.beacon.discovery.packet5_1.Header;
-import org.ethereum.beacon.discovery.packet5_1.StaticHeader;
-import org.ethereum.beacon.discovery.packet5_1.impl.HandshakeMessagePacketImpl.HandshakeAuthDataImpl;
-import org.ethereum.beacon.discovery.packet5_1.impl.OrdinaryMessageImpl.AuthDataImpl;
-import org.ethereum.beacon.discovery.packet5_1.impl.WhoAreYouPacketImpl.WhoAreYouAuthDataImpl;
+import org.ethereum.beacon.discovery.packet.AuthData;
+import org.ethereum.beacon.discovery.packet.DecodeException;
+import org.ethereum.beacon.discovery.packet.Header;
+import org.ethereum.beacon.discovery.packet.StaticHeader;
+import org.ethereum.beacon.discovery.packet.impl.HandshakeMessagePacketImpl.HandshakeAuthDataImpl;
+import org.ethereum.beacon.discovery.packet.impl.OrdinaryMessageImpl.AuthDataImpl;
+import org.ethereum.beacon.discovery.packet.impl.WhoAreYouPacketImpl.WhoAreYouAuthDataImpl;
 import org.ethereum.beacon.discovery.type.Bytes16;
 import org.ethereum.beacon.discovery.util.CryptoUtil;
