+++ src/main/java/org/ethereum/beacon/discovery/packet5_0/AbstractPacket.java
@@ -2,7 +2,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package org.ethereum.beacon.discovery.packet;
+package org.ethereum.beacon.discovery.packet5_0;
 
 import org.apache.tuweni.bytes.Bytes;
