--- src/main/java/org/ethereum/beacon/discovery/packet5_1/MessagePacket.java
@@ -1,10 +1,0 @@
-package org.ethereum.beacon.discovery.packet5_1;
-
-import org.apache.tuweni.bytes.Bytes;
-import org.ethereum.beacon.discovery.message.V5Message;
-import org.ethereum.beacon.discovery.schema.NodeRecordFactory;
-
-public interface MessagePacket<TAuthData extends AuthData> extends Packet<TAuthData> {
-
-  V5Message decryptMessage(Bytes key, NodeRecordFactory nodeRecordFactory);
-}
