+++ src/main/java/org/ethereum/beacon/discovery/pipeline/handler/UnknownPacketTagToSender.java
@@ -7,8 +7,7 @@ package org.ethereum.beacon.discovery.pipeline.handler;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.apache.tuweni.bytes.Bytes;
-import org.ethereum.beacon.discovery.packet.UnknownPacket;
-import org.ethereum.beacon.discovery.packet5_1.Packet;
+import org.ethereum.beacon.discovery.packet.Packet;
 import org.ethereum.beacon.discovery.pipeline.Envelope;
 import org.ethereum.beacon.discovery.pipeline.EnvelopeHandler;
 import org.ethereum.beacon.discovery.pipeline.Field;
