+++ src/main/java/org/ethereum/beacon/discovery/pipeline/handler/UnknownPacketTypeByStatus.java
@@ -6,9 +6,9 @@ package org.ethereum.beacon.discovery.pipeline.handler;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.ethereum.beacon.discovery.packet.AuthHeaderMessagePacket;
-import org.ethereum.beacon.discovery.packet.MessagePacket;
-import org.ethereum.beacon.discovery.packet.UnknownPacket;
+import org.ethereum.beacon.discovery.packet5_0.AuthHeaderMessagePacket;
+import org.ethereum.beacon.discovery.packet5_0.MessagePacket;
+import org.ethereum.beacon.discovery.packet5_0.UnknownPacket;
 import org.ethereum.beacon.discovery.pipeline.Envelope;
 import org.ethereum.beacon.discovery.pipeline.EnvelopeHandler;
 import org.ethereum.beacon.discovery.pipeline.Field;
