+++ src/test/java/org/ethereum/beacon/discovery/DiscoveryNetworkTest.java
@@ -17,11 +17,11 @@ import java.util.concurrent.Executors;
 import java.util.concurrent.TimeUnit;
 import org.ethereum.beacon.discovery.TestUtil.NodeInfo;
 import org.ethereum.beacon.discovery.database.Database;
-import org.ethereum.beacon.discovery.packet.AuthHeaderMessagePacket;
-import org.ethereum.beacon.discovery.packet.MessagePacket;
-import org.ethereum.beacon.discovery.packet.RandomPacket;
-import org.ethereum.beacon.discovery.packet.UnknownPacket;
-import org.ethereum.beacon.discovery.packet.WhoAreYouPacket;
+import org.ethereum.beacon.discovery.packet5_0.AuthHeaderMessagePacket;
+import org.ethereum.beacon.discovery.packet5_0.MessagePacket;
+import org.ethereum.beacon.discovery.packet5_0.RandomPacket;
+import org.ethereum.beacon.discovery.packet5_0.UnknownPacket;
+import org.ethereum.beacon.discovery.packet5_0.WhoAreYouPacket;
 import org.ethereum.beacon.discovery.scheduler.ExpirationSchedulerFactory;
 import org.ethereum.beacon.discovery.scheduler.Schedulers;
 import org.ethereum.beacon.discovery.schema.NodeRecord;
