+++ src/test/java/org/ethereum/beacon/discovery/HandshakeHandlersTest.java
@@ -29,10 +29,10 @@ import org.apache.tuweni.units.bigints.UInt64;
 import org.ethereum.beacon.discovery.TestUtil.NodeInfo;
 import org.ethereum.beacon.discovery.database.Database;
 import org.ethereum.beacon.discovery.network.NetworkParcel;
-import org.ethereum.beacon.discovery.packet.MessagePacket;
-import org.ethereum.beacon.discovery.packet.Packet;
-import org.ethereum.beacon.discovery.packet.WhoAreYouPacket;
-import org.ethereum.beacon.discovery.packet5_1.OrdinaryMessagePacket;
+import org.ethereum.beacon.discovery.packet5_0.MessagePacket;
+import org.ethereum.beacon.discovery.packet5_0.Packet;
+import org.ethereum.beacon.discovery.packet5_0.WhoAreYouPacket;
+import org.ethereum.beacon.discovery.packet.OrdinaryMessagePacket;
 import org.ethereum.beacon.discovery.pipeline.Envelope;
 import org.ethereum.beacon.discovery.pipeline.Field;
 import org.ethereum.beacon.discovery.pipeline.Pipeline;
