+++ src/test/java/org/ethereum/beacon/discovery/packet5_0/RandomPacketTest.java
@@ -1,7 +1,7 @@
 /*
  * SPDX-License-Identifier: Apache-2.0
  */
-package org.ethereum.beacon.discovery.packet;
+package org.ethereum.beacon.discovery.packet5_0;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
